refactor: passkey/WebAuthn 型・スキーマをドメインからハンドラ層へ移す
passkey はドメインの概念ではないため、中立 WebAuthn 型(webauthn-schema)を廃止し、
passkeyVerifyInputSchema と options 型付けをハンドラ層へ移す。options は client の
ceremony ライブラリ(@simplewebauthn)の型へ境界で寄せることで client の as は不要のまま。

- domain: webauthn-schema.ts を削除、auth-schema は authInputSchema のみへ、
  index.ts から passkey/WebAuthn の export を撤去
- web: passkey-schema.ts(passkeyVerifyInputSchema)を新設、passkey start ハンドラは
  @simplewebauthn の PublicKeyCredential(Creation|Request)OptionsJSON へアサーション
- verify の credential 型も @simplewebauthn の Registration/AuthenticationResponseJSON へ
- AccountUseCase の what コメントを削除
