+++ application/apps/web/src/server/auth/handler/passkey-login-verify.ts
@@ -1,10 +1,11 @@
 import { ClientError, handleError, ServerError } from '@trend-diary/common/errors'
 import { wrapAsyncCall } from '@trend-diary/common/result'
 import getRdbClient from '@trend-diary/datastore/rdb'
-import { createAccountUseCase, type PasskeyVerifyInput } from '@trend-diary/domain/user'
+import { createAccountUseCase } from '@trend-diary/domain/user'
 import { createSupabaseAuthClient } from '@/infrastructure/supabase'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
+import type { PasskeyVerifyInput } from '../passkey-schema'
 
 export default async function passkeyLoginVerify(c: ZodValidatedContext<PasskeyVerifyInput>) {
   const logger = c.get(CONTEXT_KEY.APP_LOG)
