+++ application/apps/web/src/server/auth/handler/passkey-register-start.ts
@@ -1,6 +1,6 @@
+import type { PublicKeyCredentialCreationOptionsJSON } from '@simplewebauthn/browser'
 import { handleError, ServerError } from '@trend-diary/common/errors'
 import { wrapAsyncCall } from '@trend-diary/common/result'
-import type { WebAuthnRegistrationOptions } from '@trend-diary/domain/user'
 import type { Context } from 'hono'
 import type { Env } from '@/env'
 import { createSupabaseAuthClient } from '@/infrastructure/supabase'
@@ -21,8 +21,8 @@ export default async function passkeyRegisterStart(c: Context<Env>) {
     )
   }
 
-  // Supabase SDK の options を中立な WebAuthn 型へ寄せる。hints の union が SDK 側で広い（string 拡張を含む）ぶんだけ差があるため境界で単一アサーションする
-  // oxlint-disable-next-line typescript/consistent-type-assertions -- SDK 型と中立型は hints の union 幅のみ異なる構造互換のため、境界での単一アサーションに留める
-  const options = data.options as WebAuthnRegistrationOptions
+  // Supabase SDK と @simplewebauthn の WebAuthn 型は hints の union 幅のみ異なるため、client(ceremony ライブラリ)が受ける型へ境界で寄せる
+  // oxlint-disable-next-line typescript/consistent-type-assertions -- 上記のとおり構造互換だが宣言が別のため、ライブラリ境界での単一アサーションに留める
+  const options = data.options as PublicKeyCredentialCreationOptionsJSON
   return c.json({ challengeId: data.challenge_id, options }, 200)
 }
