+++ application/apps/web/src/server/auth/handler/passkey-register-verify.ts
@@ -1,9 +1,9 @@
 import { ClientError, handleError, ServerError } from '@trend-diary/common/errors'
 import { wrapAsyncCall } from '@trend-diary/common/result'
-import type { PasskeyVerifyInput } from '@trend-diary/domain/user'
 import { createSupabaseAuthClient } from '@/infrastructure/supabase'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
+import type { PasskeyVerifyInput } from '../passkey-schema'
 
 export default async function passkeyRegisterVerify(c: ZodValidatedContext<PasskeyVerifyInput>) {
   const logger = c.get(CONTEXT_KEY.APP_LOG)
