+++ application/packages/domain/src/user/index.ts
@@ -1,29 +1,11 @@
 import { createAccountUseCase } from './factory'
 import type { ActiveUser, ActiveUserInput, CurrentUser } from './schema/active-user-schema'
 import { activeUserInputSchema, activeUserSchema } from './schema/active-user-schema'
-import type { AuthInput, PasskeyVerifyInput } from './schema/auth-schema'
-import { authInputSchema, passkeyVerifyInputSchema } from './schema/auth-schema'
-import type {
-  WebAuthnAuthenticationOptions,
-  WebAuthnRegistrationOptions,
-} from './schema/webauthn-schema'
+import type { AuthInput } from './schema/auth-schema'
+import { authInputSchema } from './schema/auth-schema'
 
 // 型
-export type {
-  ActiveUser,
-  ActiveUserInput,
-  AuthInput,
-  CurrentUser,
-  PasskeyVerifyInput,
-  WebAuthnAuthenticationOptions,
-  WebAuthnRegistrationOptions,
-}
+export type { ActiveUser, ActiveUserInput, AuthInput, CurrentUser }
 // スキーマ
 // ファクトリ
-export {
-  activeUserInputSchema,
-  activeUserSchema,
-  authInputSchema,
-  createAccountUseCase,
-  passkeyVerifyInputSchema,
-}
+export { activeUserInputSchema, activeUserSchema, authInputSchema, createAccountUseCase }
