+++ application/packages/domain/src/user/schema/auth-schema.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod'
-import type { WebAuthnCredential } from './webauthn-schema'
 
 // 認証プロバイダに依存しない入力検証。クライアントのフォームとサーバのバリデーションで共有するため domain に置く
 export const authInputSchema = z.object({
@@ -17,13 +16,3 @@ export const authInputSchema = z.object({
 })
 
 export type AuthInput = z.infer<typeof authInputSchema>
-
-// 真正性はSupabaseが検証するため中身の妥当性検証はプロバイダに委ね、ここは WebAuthn ceremony 結果を素通しする。型は中立な WebAuthn 資格情報で固定する
-export const passkeyVerifyInputSchema = z.object({
-  challengeId: z.string().min(1),
-  credential: z.custom<WebAuthnCredential>(
-    (value) => typeof value === 'object' && value !== null && !Array.isArray(value),
-  ),
-})
-
-export type PasskeyVerifyInput = z.infer<typeof passkeyVerifyInputSchema>
